Command E mods: Don't use dynamic values in require(), it confuses Webpack
Using Umzug with Webpack results in warnings on every
Webpack/webpack-dev-server build because Umzug uses require() with
dynamic values. Webpack doesn't know how to compile code when this is
done.

Example warning message

> Critical dependency: the request of a dependency is an expression

It looks like Umzug 3 is coming soon, and is a decent update. When we
upgrade to Umzug 3, we should:
* Try using it without this mod and see if Webpack throws similar
warnings.
* See if there's a way to fix this in Umzug
upstream (i.e. `noDynamicRequires: true`) although I'm not sure that's
possible because I think Webpack reads all the code, not just branches
that will be executed due to a setting.

Maybe it'd be a build-time command for Umzug, i.e.
`yarn prepare --no-dynamic-requires`.